add processMessagesInBatches

+++ domain/src/main/kotlin/dev/marcal/chatvault/app_service/dto/WppChatResponse.kt
@@ -9,7 +9,7 @@ data class ChatDTO(
     val id: Long,
     val name: String,
     val attachmentDir: String,
-    val messagesCount: Long,
+    val messagesCount: Int,
 )
 
 data class MessageDTO(
